+++ test/automated/reservation/substitute_test.rb
@@ -13,9 +13,9 @@ context "Reservation" do
       block_message = m
     end
 
-    test "Block is called" do
-      assert(block_called)
-      assert(block_message == message)
+    test "Block is not called by default" do
+      refute(block_called)
+      refute(block_message == message)
     end
 
     test "Called with message" do
@@ -24,6 +24,20 @@ context "Reservation" do
 
     test "Called with idempotence_key" do
       assert(substitute.idempotence_key?(idempotence_key))
+    end
+
+    context "When set reserved" do
+      substitute.set_reserved(true)
+
+      substitute.(message, idempotence_key) do |m|
+        block_called = true
+        block_message = m
+      end
+
+      test "Block is called" do
+        assert(block_called)
+        assert(block_message == message)
+      end
     end
   end
 end