+++ CHANGELOG.md
@@ -4,17 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.1.0 - 2023-03-29]
+### Changed
+- Don't call automatically yield in the repo, just when calling
+  set_reserved(true)
+
+
 ## [2.0.0 - 2022-12-09]
 ### Changed
-- Renamed gem and repo:
-  - from "idempotence" to "hubbado-idempotence-reservation"
+- Renamed gem and repo from "idempotence" to "hubbado-idempotence-reservation"
+
 
 ## [1.2.0 - 2022-11-29]
 ### Changed
 - Reserved metadata name is stored in metadata.local_properties
   instead of metadata.properties.
   metadata.local_properties is not followed so the info is not propagated to other messages
-  
+
 
 ## [1.1.0 - 2022-10-07]
 ### Added